Change prop to virtualTreeId and move to entrypoint
